requirements.txt
requests>=2.32.4
beautifulsoup4>=4.13.4
lxml>=6.0.0
rich-argparse>=1.7.1
python-cowsay>=1.2.1
